Rename api_methods module to api (to match name of test module).
